Weather Forecast Fixes #54.   Reorganized the tools into different personas.
