+++ backend/src/tools/finances/get_stock_financials.py
