+++ backend/src/tools/finances/get_stock_news.py
