+++ backend/src/tools/finances/get_stock_quote.py
