+++ backend/src/tools/finances/get_stock_technical_indicators.py
